refactor(admin): consolidate FE fetching, split detail page, dedup bulk
Wave 3 W3-2 (#1608):

- AUD-055: three data-fetching patterns (jsonFetch / TanStack / Refine)
  risk stale data when queryClient invalidates a jsonFetch screen.
  Add ADR-0021 (new code = useQuery), a CI guard that fails when the
  jsonFetch+useEffect file count grows, and migrate settings/locales as
  the proof slice. The remaining ~136 files are tracked in #1670.
- AUD-056: 12 of the 13 Bulk handlers shared a ~45-line copy-pasted
  skeleton (set_time_limit, chunked flush/clear, reload, complete,
  reindex). Extract AbstractBulkHandler (template method) + BulkCounters;
  each handler keeps byte-identical per-item logic (~627 lines removed).
- AUD-057: no max-lines lint let files balloon. Add a CI guard on
  files over 500 lines and split product-detail-page (1499 -> 492) into
  header/content/sidebar/other-tabs components plus data/form hooks.
  The other 16 monoliths are tracked in #1670.

Closes #1608
