+++ apps/api/src/Catalog/Application/Bulk/BulkAddCategoryHandler.php
@@ -14,7 +14,6 @@ use App\Catalog\Domain\Repository\CatalogObjectRepositoryInterface;
 use App\Catalog\Domain\Repository\ObjectCategoryRepositoryInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Uid\Uuid;
-use Throwable;
 
 /**
  * VIEW-14 (#546) — `add_category` bulk action.
@@ -22,19 +21,22 @@ use Throwable;
  * Adds the supplied category ids to each target product. Skips a
  * (product, category) pair when the assignment already exists (warning
  * log). BulkLog `old_value` snapshots the prior assignment id list to
- * support the 24h rollback path via `replaceForProduct`.
+ * support the 24h rollback path via `replaceForProduct`. Shared
+ * lifecycle: {@see AbstractBulkHandler}.
  */
-final class BulkAddCategoryHandler
+final class BulkAddCategoryHandler extends AbstractBulkHandler
 {
-    public const int CHUNK_SIZE = 200;
+    /** @var list<string> */
+    private array $categoryIds = [];
 
     public function __construct(
-        private readonly CatalogObjectRepositoryInterface $catalogObjects,
+        CatalogObjectRepositoryInterface $catalogObjects,
         private readonly ObjectCategoryRepositoryInterface $objectCategories,
-        private readonly EntityManagerInterface $em,
-        private readonly BulkContext $bulkContext,
-        private readonly BulkReindexQueueInterface $reindexQueue,
+        EntityManagerInterface $em,
+        BulkContext $bulkContext,
+        BulkReindexQueueInterface $reindexQueue,
     ) {
+        parent::__construct($catalogObjects, $em, $bulkContext, $reindexQueue);
     }
 
     /**
@@ -44,119 +46,57 @@ final class BulkAddCategoryHandler
      */
     public function handle(BulkSession $session, array $categoryIds): array
     {
-        // Long bulk runs (2k+ products) routinely exceed PHP's
-        // 30s HTTP timeout — without disabling it the handler
-        // is killed mid-loop, the session stays incomplete, and
-        // the operator sees 200 + zero rows touched.
-        set_time_limit(0);
+        $this->categoryIds = $categoryIds;
 
-        $this->bulkContext->setBulk(true, $session->getId());
-        try {
-            $success = 0;
-            $skipped = 0;
-            $errors = 0;
-            $chunkCounter = 0;
+        return $this->runBatch($session);
+    }
 
-            foreach ($session->getTargetObjectIds() as $targetId) {
-                try {
-                    $product = $this->catalogObjects->findById(Uuid::fromString($targetId));
-                    if (!$product instanceof CatalogObject) {
-                        ++$errors;
-                        ++$chunkCounter;
-                        continue;
-                    }
+    protected function processObject(CatalogObject $object, BulkSession $session, BulkCounters $counters): void
+    {
+        $existing = $this->objectCategories->findByProduct($object);
+        $existingIds = array_map(
+            static fn (ObjectCategory $oc): string => $oc->getCategory()->getId()->toRfc4122(),
+            $existing,
+        );
 
-                    $existing = $this->objectCategories->findByProduct($product);
-                    $existingIds = array_map(
-                        static fn (ObjectCategory $oc): string => $oc->getCategory()->getId()->toRfc4122(),
-                        $existing,
-                    );
-
-                    $added = 0;
-                    foreach ($categoryIds as $catId) {
-                        if (\in_array($catId, $existingIds, true)) {
-                            continue;
-                        }
-                        $category = $this->catalogObjects->findById(Uuid::fromString($catId));
-                        if (!$category instanceof CatalogObject) {
-                            continue;
-                        }
-                        $this->objectCategories->save(new ObjectCategory($product, $category));
-                        ++$added;
-                    }
-
-                    if (0 === $added) {
-                        ++$skipped;
-                        $this->em->persist(new BulkLog(
-                            $session->getId(),
-                            $product->getId(),
-                            null,
-                            $existingIds,
-                            $existingIds,
-                            BulkLog::LEVEL_WARNING,
-                            'All categories already assigned',
-                        ));
-                    } else {
-                        $product->markTouchedByBulkSession($session->getId());
-                        $this->em->persist(new BulkLog(
-                            $session->getId(),
-                            $product->getId(),
-                            null,
-                            $existingIds,
-                            array_values(array_unique([...$existingIds, ...$categoryIds])),
-                            BulkLog::LEVEL_INFO,
-                            null,
-                        ));
-                        ++$success;
-                    }
-                } catch (Throwable $e) {
-                    ++$errors;
-                    $this->em->persist(new BulkLog(
-                        $session->getId(),
-                        Uuid::fromString($targetId),
-                        null,
-                        null,
-                        null,
-                        BulkLog::LEVEL_ERROR,
-                        $e->getMessage(),
-                    ));
-                }
-
-                ++$chunkCounter;
-                if ($chunkCounter >= self::CHUNK_SIZE) {
-                    $this->em->flush();
-                    $this->em->clear();
-                    $chunkCounter = 0;
-                }
+        $added = 0;
+        foreach ($this->categoryIds as $catId) {
+            if (\in_array($catId, $existingIds, true)) {
+                continue;
             }
-
-            if ($chunkCounter > 0) {
-                $this->em->flush();
+            $category = $this->catalogObjects->findById(Uuid::fromString($catId));
+            if (!$category instanceof CatalogObject) {
+                continue;
             }
-
-            // Reload BulkSession: per-chunk em->clear() detached the
-
-            // local instance and its Tenant proxy. The final persist
-
-            // below would otherwise raise EntityNotFoundException on
-
-            // flush trying to resolve the stale proxy.
-
-            $reloaded = $this->em->find(BulkSession::class, $session->getId());
-
-            if ($reloaded instanceof BulkSession) {
-                $session = $reloaded;
-            }
-
-            $session->complete($success, $skipped, $errors);
-            $this->em->persist($session);
-            $this->em->flush();
-
-            $this->reindexQueue->queueAll($session->getTargetObjectIds());
-
-            return ['success' => $success, 'skipped' => $skipped, 'error' => $errors];
-        } finally {
-            $this->bulkContext->setBulk(false);
+            $this->objectCategories->save(new ObjectCategory($object, $category));
+            ++$added;
         }
+
+        if (0 === $added) {
+            ++$counters->skipped;
+            $this->em->persist(new BulkLog(
+                $session->getId(),
+                $object->getId(),
+                null,
+                $existingIds,
+                $existingIds,
+                BulkLog::LEVEL_WARNING,
+                'All categories already assigned',
+            ));
+
+            return;
+        }
+
+        $object->markTouchedByBulkSession($session->getId());
+        $this->em->persist(new BulkLog(
+            $session->getId(),
+            $object->getId(),
+            null,
+            $existingIds,
+            array_values(array_unique([...$existingIds, ...$this->categoryIds])),
+            BulkLog::LEVEL_INFO,
+            null,
+        ));
+        ++$counters->success;
     }
 }
