+++ apps/api/src/Catalog/Application/Bulk/BulkIncrementNumericHandler.php
@@ -13,8 +13,6 @@ use App\Catalog\Domain\Entity\CatalogObject;
 use App\Catalog\Domain\Repository\CatalogObjectRepositoryInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
-use Symfony\Component\Uid\Uuid;
-use Throwable;
 
 /**
  * VIEW-13 (#545) — `increment_numeric` bulk action.
@@ -25,19 +23,22 @@ use Throwable;
  * Locked attributes (VIEW-33 / PRD §8.3) skip with a warning entry.
  *
  * Division-by-zero raises a per-row error log (no-op) rather than
- * aborting the batch.
+ * aborting the batch. Shared lifecycle: {@see AbstractBulkHandler}.
  */
-final class BulkIncrementNumericHandler
+final class BulkIncrementNumericHandler extends AbstractBulkHandler
 {
-    public const int CHUNK_SIZE = 200;
+    private string $attrCode = '';
+    private string $operator = '+';
+    private float $operand = 0.0;
 
     public function __construct(
-        private readonly CatalogObjectRepositoryInterface $catalogObjects,
-        private readonly EntityManagerInterface $em,
-        private readonly BulkContext $bulkContext,
+        CatalogObjectRepositoryInterface $catalogObjects,
+        EntityManagerInterface $em,
+        BulkContext $bulkContext,
         private readonly AttributeLockReader $lockReader,
-        private readonly BulkReindexQueueInterface $reindexQueue,
+        BulkReindexQueueInterface $reindexQueue,
     ) {
+        parent::__construct($catalogObjects, $em, $bulkContext, $reindexQueue);
     }
 
     /**
@@ -45,154 +46,90 @@ final class BulkIncrementNumericHandler
      */
     public function handle(BulkSession $session, string $attrCode, string $operator, float $operand): array
     {
-        // Long bulk runs (2k+ products) routinely exceed PHP's 30s HTTP
-        // timeout — without disabling it the handler is killed mid-loop,
-        // the session stays incomplete, and the operator sees 200 + zero
-        // rows touched.
-        set_time_limit(0);
-
         if (!\in_array($operator, ['+', '-', '*', '/', '%'], true)) {
             throw new BadRequestHttpException(\sprintf('Unsupported operator "%s".', $operator));
         }
 
-        $this->bulkContext->setBulk(true, $session->getId());
-        try {
-            $success = 0;
-            $skipped = 0;
-            $errors = 0;
-            $chunkCounter = 0;
+        $this->attrCode = $attrCode;
+        $this->operator = $operator;
+        $this->operand = $operand;
 
-            foreach ($session->getTargetObjectIds() as $targetId) {
-                try {
-                    $object = $this->catalogObjects->findById(Uuid::fromString($targetId));
-                    if (!$object instanceof CatalogObject) {
-                        ++$errors;
-                        ++$chunkCounter;
-                        continue;
-                    }
+        return $this->runBatch($session);
+    }
 
-                    if ($this->lockReader->isLocked($object, $attrCode)) {
-                        ++$skipped;
-                        $this->em->persist(new BulkLog(
-                            $session->getId(),
-                            $object->getId(),
-                            null,
-                            $object->getAttributesIndexed()[$attrCode] ?? null,
-                            $object->getAttributesIndexed()[$attrCode] ?? null,
-                            BulkLog::LEVEL_WARNING,
-                            'Attribute locked',
-                        ));
-                        ++$chunkCounter;
-                        if ($chunkCounter >= self::CHUNK_SIZE) {
-                            $this->em->flush();
-                            $this->em->clear();
-                            $chunkCounter = 0;
-                        }
-                        continue;
-                    }
+    protected function processObject(CatalogObject $object, BulkSession $session, BulkCounters $counters): void
+    {
+        if ($this->lockReader->isLocked($object, $this->attrCode)) {
+            ++$counters->skipped;
+            $this->em->persist(new BulkLog(
+                $session->getId(),
+                $object->getId(),
+                null,
+                $object->getAttributesIndexed()[$this->attrCode] ?? null,
+                $object->getAttributesIndexed()[$this->attrCode] ?? null,
+                BulkLog::LEVEL_WARNING,
+                'Attribute locked',
+            ));
 
-                    $indexed = $object->getAttributesIndexed();
-                    $oldValue = $indexed[$attrCode] ?? null;
-
-                    if (!is_numeric($oldValue)) {
-                        ++$skipped;
-                        $this->em->persist(new BulkLog(
-                            $session->getId(),
-                            $object->getId(),
-                            null,
-                            $oldValue,
-                            $oldValue,
-                            BulkLog::LEVEL_WARNING,
-                            'Value is not numeric',
-                        ));
-                        ++$chunkCounter;
-                        continue;
-                    }
-
-                    $current = (float) $oldValue;
-                    $newValue = match ($operator) {
-                        '+' => $current + $operand,
-                        '-' => $current - $operand,
-                        '*' => $current * $operand,
-                        '/' => $this->safeDivide($current, $operand),
-                        '%' => $this->safeModulo($current, $operand),
-                    };
-
-                    if (null === $newValue) {
-                        ++$errors;
-                        $this->em->persist(new BulkLog(
-                            $session->getId(),
-                            $object->getId(),
-                            null,
-                            $oldValue,
-                            null,
-                            BulkLog::LEVEL_ERROR,
-                            'Division by zero',
-                        ));
-                    } else {
-                        $indexed[$attrCode] = $newValue;
-                        $object->updateAttributeIndex($indexed);
-                        $object->markTouchedByBulkSession($session->getId());
-                        $this->em->persist(new BulkLog(
-                            $session->getId(),
-                            $object->getId(),
-                            null,
-                            $oldValue,
-                            $newValue,
-                            BulkLog::LEVEL_INFO,
-                            null,
-                        ));
-                        ++$success;
-                    }
-                } catch (Throwable $e) {
-                    ++$errors;
-                    $this->em->persist(new BulkLog(
-                        $session->getId(),
-                        Uuid::fromString($targetId),
-                        null,
-                        null,
-                        null,
-                        BulkLog::LEVEL_ERROR,
-                        $e->getMessage(),
-                    ));
-                }
-
-                ++$chunkCounter;
-                if ($chunkCounter >= self::CHUNK_SIZE) {
-                    $this->em->flush();
-                    $this->em->clear();
-                    $chunkCounter = 0;
-                }
-            }
-
-            if ($chunkCounter > 0) {
-                $this->em->flush();
-            }
-
-            // Reload BulkSession: per-chunk em->clear() detached the
-
-            // local instance and its Tenant proxy. The final persist
-
-            // below would otherwise raise EntityNotFoundException on
-
-            // flush trying to resolve the stale proxy.
-
-            $reloaded = $this->em->find(BulkSession::class, $session->getId());
-
-            if ($reloaded instanceof BulkSession) {
-                $session = $reloaded;
-            }
-
-            $session->complete($success, $skipped, $errors);
-            $this->em->persist($session);
-            $this->em->flush();
-
-            $this->reindexQueue->queueAll($session->getTargetObjectIds());
-
-            return ['success' => $success, 'skipped' => $skipped, 'error' => $errors];
-        } finally {
-            $this->bulkContext->setBulk(false);
+            return;
         }
+
+        $indexed = $object->getAttributesIndexed();
+        $oldValue = $indexed[$this->attrCode] ?? null;
+
+        if (!is_numeric($oldValue)) {
+            ++$counters->skipped;
+            $this->em->persist(new BulkLog(
+                $session->getId(),
+                $object->getId(),
+                null,
+                $oldValue,
+                $oldValue,
+                BulkLog::LEVEL_WARNING,
+                'Value is not numeric',
+            ));
+
+            return;
+        }
+
+        $current = (float) $oldValue;
+        $newValue = match ($this->operator) {
+            '+' => $current + $this->operand,
+            '-' => $current - $this->operand,
+            '*' => $current * $this->operand,
+            '/' => $this->safeDivide($current, $this->operand),
+            '%' => $this->safeModulo($current, $this->operand),
+            default => null,
+        };
+
+        if (null === $newValue) {
+            ++$counters->error;
+            $this->em->persist(new BulkLog(
+                $session->getId(),
+                $object->getId(),
+                null,
+                $oldValue,
+                null,
+                BulkLog::LEVEL_ERROR,
+                'Division by zero',
+            ));
+
+            return;
+        }
+
+        $indexed[$this->attrCode] = $newValue;
+        $object->updateAttributeIndex($indexed);
+        $object->markTouchedByBulkSession($session->getId());
+        $this->em->persist(new BulkLog(
+            $session->getId(),
+            $object->getId(),
+            null,
+            $oldValue,
+            $newValue,
+            BulkLog::LEVEL_INFO,
+            null,
+        ));
+        ++$counters->success;
     }
 
     private function safeDivide(float $a, float $b): ?float
