+++ apps/api/src/Catalog/Application/Bulk/BulkMoveCategoryHandler.php
@@ -14,7 +14,6 @@ use App\Catalog\Domain\Repository\CatalogObjectRepositoryInterface;
 use App\Catalog\Domain\Repository\ObjectCategoryRepositoryInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Uid\Uuid;
-use Throwable;
 
 /**
  * VIEW-14 (#546) — `move_category` bulk action.
@@ -22,19 +21,25 @@ use Throwable;
  * Full-replace assignment list. Wipes existing rows and re-inserts the
  * supplied target ids in a single transaction per product via the
  * repository's `replaceForProduct`. Used by the *„Przenieś kategorię"*
- * mockup flow when the operator picks a fresh taxonomy slot.
+ * mockup flow when the operator picks a fresh taxonomy slot. Shared
+ * lifecycle: {@see AbstractBulkHandler}.
  */
-final class BulkMoveCategoryHandler
+final class BulkMoveCategoryHandler extends AbstractBulkHandler
 {
-    public const int CHUNK_SIZE = 200;
+    /** @var list<string> */
+    private array $categoryIds = [];
+    private ?Uuid $primaryId = null;
+    /** @var list<Uuid> */
+    private array $targetUuids = [];
 
     public function __construct(
-        private readonly CatalogObjectRepositoryInterface $catalogObjects,
+        CatalogObjectRepositoryInterface $catalogObjects,
         private readonly ObjectCategoryRepositoryInterface $objectCategories,
-        private readonly EntityManagerInterface $em,
-        private readonly BulkContext $bulkContext,
-        private readonly BulkReindexQueueInterface $reindexQueue,
+        EntityManagerInterface $em,
+        BulkContext $bulkContext,
+        BulkReindexQueueInterface $reindexQueue,
     ) {
+        parent::__construct($catalogObjects, $em, $bulkContext, $reindexQueue);
     }
 
     /**
@@ -44,96 +49,33 @@ final class BulkMoveCategoryHandler
      */
     public function handle(BulkSession $session, array $categoryIds): array
     {
-        // Long bulk runs (2k+ products) routinely exceed PHP's
-        // 30s HTTP timeout — without disabling it the handler
-        // is killed mid-loop, the session stays incomplete, and
-        // the operator sees 200 + zero rows touched.
-        set_time_limit(0);
+        $this->categoryIds = $categoryIds;
+        $this->primaryId = [] === $categoryIds ? null : Uuid::fromString($categoryIds[0]);
+        $this->targetUuids = array_map(static fn (string $id) => Uuid::fromString($id), $categoryIds);
 
-        $this->bulkContext->setBulk(true, $session->getId());
-        try {
-            $success = 0;
-            $errors = 0;
-            $chunkCounter = 0;
-            $primaryId = [] === $categoryIds ? null : Uuid::fromString($categoryIds[0]);
-            $targetUuids = array_map(static fn (string $id) => Uuid::fromString($id), $categoryIds);
+        return $this->runBatch($session);
+    }
 
-            foreach ($session->getTargetObjectIds() as $targetId) {
-                try {
-                    $product = $this->catalogObjects->findById(Uuid::fromString($targetId));
-                    if (!$product instanceof CatalogObject) {
-                        ++$errors;
-                        ++$chunkCounter;
-                        continue;
-                    }
+    protected function processObject(CatalogObject $object, BulkSession $session, BulkCounters $counters): void
+    {
+        $existing = $this->objectCategories->findByProduct($object);
+        $existingIds = array_map(
+            static fn (ObjectCategory $oc): string => $oc->getCategory()->getId()->toRfc4122(),
+            $existing,
+        );
 
-                    $existing = $this->objectCategories->findByProduct($product);
-                    $existingIds = array_map(
-                        static fn (ObjectCategory $oc): string => $oc->getCategory()->getId()->toRfc4122(),
-                        $existing,
-                    );
+        $this->objectCategories->replaceForProduct($object, $this->targetUuids, $this->primaryId);
+        $object->markTouchedByBulkSession($session->getId());
 
-                    $this->objectCategories->replaceForProduct($product, $targetUuids, $primaryId);
-                    $product->markTouchedByBulkSession($session->getId());
-
-                    $this->em->persist(new BulkLog(
-                        $session->getId(),
-                        $product->getId(),
-                        null,
-                        $existingIds,
-                        $categoryIds,
-                        BulkLog::LEVEL_INFO,
-                        null,
-                    ));
-                    ++$success;
-                } catch (Throwable $e) {
-                    ++$errors;
-                    $this->em->persist(new BulkLog(
-                        $session->getId(),
-                        Uuid::fromString($targetId),
-                        null,
-                        null,
-                        null,
-                        BulkLog::LEVEL_ERROR,
-                        $e->getMessage(),
-                    ));
-                }
-
-                ++$chunkCounter;
-                if ($chunkCounter >= self::CHUNK_SIZE) {
-                    $this->em->flush();
-                    $this->em->clear();
-                    $chunkCounter = 0;
-                }
-            }
-
-            if ($chunkCounter > 0) {
-                $this->em->flush();
-            }
-
-            // Reload BulkSession: per-chunk em->clear() detached the
-
-            // local instance and its Tenant proxy. The final persist
-
-            // below would otherwise raise EntityNotFoundException on
-
-            // flush trying to resolve the stale proxy.
-
-            $reloaded = $this->em->find(BulkSession::class, $session->getId());
-
-            if ($reloaded instanceof BulkSession) {
-                $session = $reloaded;
-            }
-
-            $session->complete($success, 0, $errors);
-            $this->em->persist($session);
-            $this->em->flush();
-
-            $this->reindexQueue->queueAll($session->getTargetObjectIds());
-
-            return ['success' => $success, 'skipped' => 0, 'error' => $errors];
-        } finally {
-            $this->bulkContext->setBulk(false);
-        }
+        $this->em->persist(new BulkLog(
+            $session->getId(),
+            $object->getId(),
+            null,
+            $existingIds,
+            $this->categoryIds,
+            BulkLog::LEVEL_INFO,
+            null,
+        ));
+        ++$counters->success;
     }
 }
